moves odb listeners / msg fetching to App state level loads history :)
