+++ frontend/src/components/pages/MailboxScreen.tsx
@@ -1,44 +1,13 @@
 import { VStack } from '@chakra-ui/react';
 import MailboxItem from 'components/molecules/MailboxItem';
-import OrbitDbTest from 'components/organisms/OrbitDbTest';
-import { State, StoredMessage } from 'contexts/State';
+import { State } from 'contexts/State';
 
 type MailboxScreenProps = {
   state: State;
   openMessage: (id: string) => void;
-  addMessage: (msg: StoredMessage) => void;
 };
 
 function MailboxScreen(props: MailboxScreenProps) {
-  const { addMessage } = props;
-
-  // const { odb, db, setDbName } = useOrbitDb();
-
-  // useEffect(() => {
-  //   //todo: set to app states' did
-  //   setDbName('did:key:z6MkvpyzVtYLETJFaYXs3My4vtXMdZs7SjmmQQEpL9nH7MmY')
-  // },[setDbName]);
-
-  // useEffect(() => {
-  //   if (!db) return;
-  //   db.events.on('replicated', (address: string) => {
-  //     const all = db
-  //       .iterator({ limit: 10 })
-  //       .collect()
-  //       .reverse()
-  //       .map((e: any) => e.payload.value);
-  //     const latest = all[0]; //<-- a really bad hack.
-  //     // const newMsg = {
-  //     //   id: ;
-  //     //   date: number;
-  //     //   sender: string;
-  //     //   status: MessageSavingStatus;
-  //     //   subject: string;
-  //     //   content: string;
-  //     // }
-  //   });
-  // }, [db]);
-
   return (
     <VStack align="stretch" spacing={0} width="inherit">
       {props.state.messages.map((message, idx) => (
@@ -48,7 +17,6 @@ function MailboxScreen(props: MailboxScreenProps) {
           onClick={() => props.openMessage(message.id)}
         />
       ))}
-      <OrbitDbTest addMessage={addMessage} />
     </VStack>
   );
 }
