+++ frontend/src/contexts/OrbitDB.tsx
@@ -60,10 +60,8 @@ const OrbitDbProvider = ({
         console.debug('start loading', dbname)
       );
       db.events.on('ready', (dbname: string) => console.log('ready', dbname));
-
-      
+      console.log("start loading odb");
       await db.load();
-
       setDb(db);
     })();
   }, [odb, ipfs, dbName]);
